# refactor(core): internalize, simplify and optimize the SSG logic