--- a/packages/docusaurus/src/client/serverEntry.tsx
+++ b/packages/docusaurus/src/client/serverEntry.tsx
@@ -18,31 +18,7 @@ import {
 } from './BrokenLinksContext';
 import type {PageCollectedData, AppRenderer} from '../common';
 
-function buildSSRErrorMessage({
-  error,
-  pathname,
-}: {
-  error: Error;
-  pathname: string;
-}): string {
-  const parts = [
-    `Docusaurus server-side rendering could not render static page with path ${pathname} because of error: ${error.message}`,
-  ];
-
-  const isNotDefinedErrorRegex =
-    /(?:window|document|localStorage|navigator|alert|location|buffer|self) is not defined/i;
-
-  if (isNotDefinedErrorRegex.test(error.message)) {
-    // prettier-ignore
-    parts.push(`It looks like you are using code that should run on the client-side only.
-To get around it, try using \`<BrowserOnly>\` (https://docusaurus.io/docs/docusaurus-core/#browseronly) or \`ExecutionEnvironment\` (https://docusaurus.io/docs/docusaurus-core/#executionenvironment).
-It might also require to wrap your client code in \`useEffect\` hook and/or import a third-party library dynamically (if any).`);
-  }
-
-  return parts.join('\n');
-}
-
-const doRender: AppRenderer = async ({pathname}) => {
+const render: AppRenderer = async ({pathname}) => {
   await preload(pathname);
 
   const modules = new Set<string>();
@@ -64,26 +40,15 @@ const doRender: AppRenderer = async ({pathname}) => {
   );
 
   const html = await renderToHtml(app);
-  const {helmet} = helmetContext as FilledContext;
 
   const collectedData: PageCollectedData = {
+    headTags: (helmetContext as FilledContext).helmet,
     anchors: statefulBrokenLinks.getCollectedAnchors(),
     links: statefulBrokenLinks.getCollectedLinks(),
-    headTags: helmet,
     modules: Array.from(modules),
   };
 
   return {html, collectedData};
-};
-
-const render: AppRenderer = async (params) => {
-  try {
-    return await doRender(params);
-  } catch (errorUnknown) {
-    const error = errorUnknown as Error;
-    const message = buildSSRErrorMessage({error, pathname: params.pathname});
-    throw new Error(message, {cause: error});
-  }
 };
 
 export default render;