… packages/docusaurus/package.json
… packages/docusaurus/src/commands/deploy.ts
… packages/docusaurus/src/commands/serve.ts
… packages/docusaurus/src/commands/start.ts
… packages/docusaurus/src/common.d.ts
… packages/docusaurus/src/deps.d.ts
… packages/docusaurus/src/ssg.ts
… packages/docusaurus/src/templates/dev.html.template.ejs
… packages/docusaurus/src/templates/ssr.html.template.ts
… packages/docusaurus/src/templates/templates.ts
… packages/docusaurus/src/types.d.ts
… packages/docusaurus/src/utils.ts
… packages/docusaurus/src/webpack/__tests__/base.test.ts
… packages/docusaurus/src/webpack/__tests__/client.test.ts
… packages/docusaurus/src/webpack/__tests__/server.test.ts
… packages/docusaurus/src/webpack/base.ts
… packages/docusaurus/src/webpack/client.ts
… packages/docusaurus/src/webpack/minification.ts
… packages/docusaurus/src/webpack/server.ts
… packages/docusaurus/src/webpack/ssg.ts
… packages/docusaurus/src/webpack/utils.ts
… project-words.txt
… yarn.lock