## Motivation

Our static site generation logic initially comes from an unmaintained Webpack plugin, that we forked:
- Original project: https://github.com/markdalgleish/static-site-generator-webpack-plugin
- Our 2022-2024: https://github.com/slorber/static-site-generator-webpack-plugin

It's time to re-internalize this code, remove some unused code, and remove friction to be able to more easily optimize it.

This:
- decouples the SSG pipeline from webpack
- refactor/clean all webpack configs and start/build scripts
- add a basic perf logging system (internal for now)
- optimize SSG pipeline
- reduce server bundle size, moving dependencies (eta, minifier) out of the server bundle



The build time perf wins are not huge, because SSG executes relatively fast compared to the bundling phase.

```bash
yarn clear && yarn install && yarn build:website --locale en && hyperfine --runs 5 'yarn build:website --locale en'

Before
  Time (mean ± σ):     39.741 s ±  1.187 s    [User: 44.686 s, System: 6.152 s]
  Range (min … max):   38.486 s … 41.364 s    10 runs


After
  Time (mean ± σ):     37.416 s ±  1.405 s    [User: 42.558 s, System: 6.299 s]
  Range (min … max):   36.015 s … 39.096 s    5 runs
```

## Test Plan

CI + unit tests

### Test links

https://deploy-preview-9798--docusaurus-2.netlify.app/